[Justice Counts] Convert MetricInterface objects to storage json representation. (Recidiviz/recidiviz-data#28439)
## Description of the change

Add the to_storage_json method to convert MetricInterfaces to the json
object which will be stored in the MetricSetting table.

to_storage_json() jsonifies each field in the MetricInterface, but drops
all report datapoints (which should be stored in the datapoints table
instead) as well as many of the Null/None metric interface fields (so
that the json representation is lighter. The Null/None metric interface
fields will be re-populated according to the MetricDefinition later when
we read the object from the table.

Zenhub Ticket Recidiviz/recidiviz-data#28420

## Testing
Added multiple unit tests for a variety of MetricInterfaces.

GitOrigin-RevId: e03a0c1a2284819757e7e255327bbdb41c98293e