[US_ND] Refactor state staff view to pull most recent officer name (Recidiviz/recidiviz-data#28483)
## Description of the change

Previously, the `docstars_staff` ingest view that hydrates `state_staff`
in ND was relying on the manual roster as a primary source of truth, and
only supplementing from nightly transfers of raw data those officers who
did not appear at all in the manual roster. We ran into an issue where
officers have changed their names since we last received a roster, and
were not seeing their correct names in the ET workflows tool. This would
likely be fixed in the next roster transfer, but this change makes the
view more resilient to these kinds of intramonth changes.

Instead of relying exclusively on the roster for the most up-to-date
information when an officer also appears in nightly transfers, we now
compare the two sources and choose the information that is the most
recent. There are three scenarios to gauge the impact of this change:
- If a person appears in both data sources (the monthly roster and the
nightly transfers), with the same information in both, this will have no
impact.
- If a person appears in both data sources, but has had a recent name
change, this will get their correct information into the view faster.
- If a person only appears in one data source, this will have no impact.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-dashboardsRecidiviz/recidiviz-data#5167

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 544655593bb444078676464d5c83704d03f1b79f