[Insights] Separate existing outliers_staff_count_pct_change into two distinct validations with diff thresholds (Recidiviz/recidiviz-data#28441)
## Description of the change

## ☑️ ready for review

Updated validation so that we now have 2 separate percent change
validations:
1. one that has a zero threshold for percent change within a month (and
passes on the first of every month)
2. one that has the default 2% threshold for percent change between the
last day of the previous month and the first day of the current month

This change was needed to reduce noise and have meaningful validation
failures since the original validation would always fail on the first of
the month when we change the period that is surfaced in Insights
products. See the validation view descriptions for a detailed breakdown
of the differences between the two.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#28238

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: aef6e1f0e958a32d0415addc2b6f833b45add798