[Justice Counts] New endpoint for fetching child agency information (Recidiviz/recidiviz-data#28492)
## Description of the change

For superagencies, we want to:
1) Add a table on the home page showing them a list of their child
agencies
2) Allow them to easily toggle between all of their child agencies on
the Metric Settings and Explore Data pages

Right now, the FE has no easy way of knowing all of the child agencies
belonging to a superagency. In
https://github.com/Recidiviz/recidiviz-data/pull/28156, I added this
info to the endpoint that powers the home page, which enables use case 1
but not 2. Here I'm also creating a new endpoint that can be called from
any page. As per @mxosman 's comment
[here](https://github.com/Recidiviz/recidiviz-data/pull/28488#issuecomment-2024092841),
we will then store the results of this fetch so they are still
accessible if the user switches pages.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: f438bdfcd266622824f8b681ad41fabc271d41b7