[Justice Counts] [Publisher Notifications] auto-populate offset to be 15 when un-configured (Recidiviz/recidiviz-data#28521)
## Description of the change

If a user has not set an offset, it will auto-populate to 15. This
change came up as a request from playtesting from Jessica!

https://github.com/Recidiviz/recidiviz-data/assets/19961693/8a70d803-5656-441c-a043-f29dfff494ef

## Related issues

Closes Recidiviz/recidiviz-data#28519

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 4a2016ae14661b882c8920af3aa19ae9f3f28f76