[US_OR] Changing table used for supervision level in Supervision Periods (Recidiviz/recidiviz-data#28335)
## Description of the change

The original table OR sent for supervision level seemed to not be
reliable and during further discussions, this new table came up. This
change is switching to use that table instead.

Periods with null supervision levels is currently around 68% and these
changes bring it to about 10%. Currently 11383 open periods have a null
supervision level and this change brings that number to 2480.

The number of overall periods increases from  2,146,984 to 2,436,288.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#28131
Relates to Recidiviz/recidiviz-data#26891

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: bc7a73901986392311330bb42b98c6eb10a254a8