[Dataflow] Do not identify events that are for skipped metrics (Recidiviz/recidiviz-data#28455)
## Description of the change

Our metric pipelines have always had logic to filter which metrics are
produced, via the `metric_types` parameter. However, we would still
identify all events that are needed for any metric type, which is the
expensive part of the process.

This updates the filtering logic to include the identifier, and only
identify events that are necessary for the specified metric types,
skipping any logic that generates events for metrics that are skipped.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

part of Recidiviz/recidiviz-data#9266

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 04b7882e7fbcd94c0d5d098dd1d92ef9b11ffce8