[Documentation] Prune output of descendants_dfs_tree_str()/ancestors_dfs_tree_str() (Recidiviz/recidiviz-data#28534)
## Description of the change

The `recidiviz.tools.display_bq_dag_for_view` script has a runtime that
scales with the number of *paths* from the selected view to all leaf
nodes in the BQ graph (or root nodes, depending on the dependency tree
direction).

Due to the highly connected nature of our graph (e.g. with views like
person_events that most views feed into), this script has basically
become unusable for many views because it printed out pages and pages of
redundant dependency info.

Consider the following graph:
```
   A
 /   \
B     C
 \   /
   D
   |
<huge subgraph>
```
In the old world we'd print out the dependencies for the "huge subgraph"
twice, once for the path through B, then again for the path through C.

In Recidiviz/recidiviz-data#28260, Mayuka introduced a view change that significantly increases
the number of paths from `dataflow_sessions` to the edge of our view
graph, which was timing out one of our tests. It also indicates that the
change would render that script fully unusable for many views.

In this PR, I reimplemented our tree printing code to truncate the
printout for the second time we get to a subgraph we've already fully
traversed.

In order to review this I would spot check some of the markdown changes
to see how the new docstrings look, then look at the test behavior, then
actually dive into the code. Sorry it's a lot at once!

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Unblocks Recidiviz/recidiviz-data#28260.

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: d69e5e912f280a6a074c41151b750e354d3acc94