[Justice Counts] [Publisher Notifications] Temporarily send all reminder emails send in Production Nichelle and Lili (Recidiviz/recidiviz-data#28548)
## Description of the change

I'm launching Publisher Notifications next week! For the first round of
emails I'm going to send all emails to Lili and I so that we keep a
close eye on the project and make sure everything is running smoothly.

I tested with dry run to make sure that the script would send the emails
to Lili and I!

I also checked with Peter to be sure that it was safe to explicitly have
Lili's and my emails in the ode base. This is how he answered:

```
think so

As always, be careful for phishing attempts

but I don't think that is a problem especially since you have Google Advanced Protection enabled
```

## Related issues

Closes Recidiviz/recidiviz-data#28546

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 0a52965420ba10e9c222259054052122d1a5a6cf