[US_TN] Fix CAF Q7 Notes to use incident class to determine rank of most serious disciplinary (Recidiviz/recidiviz-data#28591)
## Description of the change

For the most serious disciplinary question on the CAF form, the metadata
is supposed to show (all) the most serious disciplinaries in the last 18
months. Because of a bug in the field that was being used to select the
`RANK()`, this was not happening. This PR uses `incident_class` instead
of `note_body` to determine which incidents to display.

Validation notes:
1. Number of clients in total in `us_tn_custody_level_downgrade_record`
is unchanged: 706 records in both sandbox and non sandbox versions
```
select count(*)
  from `recidiviz-123.dsharm20240401_workflows_views.us_tn_custody_level_downgrade_record_materialized`
```
2. The query below shows that currently, there are 167 clients for whom
the Q7 notes are not unique on incident class (e.g. notes include Class
As and Bs in the last 18 months, which should not happen) and in the
sandbox version we dont see this
```
SELECT countif(min_class != max_class),
       count(distinct(if(min_class != max_class,external_id,NULL )))
FROM (
  SELECT
      external_id,
      MIN(most_serious_class) OVER(PARTITION BY external_id) AS min_class,
      MAX(most_serious_class) OVER(PARTITION BY external_id) AS max_class,
FROM (
  select external_id,
        SPLIT(JSON_EXTRACT_SCALAR(form_information_q7_notes_2,'$.note_body'),' ')[OFFSET(1)] AS most_serious_class
  from `recidiviz-123.workflows_views.us_tn_custody_level_downgrade_record_materialized`,
  UNNEST(JSON_QUERY_ARRAY(form_information_q7_notes)) AS form_information_q7_notes_2
  WHERE ARRAY_LENGTH(JSON_QUERY_ARRAY(form_information_q7_notes)) != 1
)

)
```

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: e701fe935ce9ff30a9e1227abf7dcaa75d5c40cf