[Workflows][US_OR] Updates OR ED record view to add client birthdate (Recidiviz/recidiviz-data#28709)
## Description of the change

This ticket is part of the larger
[epic](https://app.zenhub.com/workspaces/polaris-614c9f27fa19b10015ddd495/issues/gh/recidiviz/recidiviz-data/28100)
of requests for OR Earned Discharge updates.

In particular, this PR adds the client birthdate as a form information
column on the `us_or_earned_discharge_record` view. Frontend change to
follow.

Testing:
Ran the `load_views_to_sandbox` script to confirm the column was added
to the view. Not sure if there are other typical testing methods🤔 I
think the `compare_views` script wasn't applicable since there's an
additional column in the new view.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Recidiviz/recidiviz-data#28125

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: d475e6455be25faebc058e9b95346a135e0e1928